Move the lesson sel and add unit sel
