+++ apps/src/templates/teacherDashboardShared/lessonSelector.module.scss
@@ -1,0 +1,5 @@
+.skeletonDropdown {
+  height: 32px;
+  width: 200px;
+  border-radius: 4px;
+}
